smurf: Kluge STACKFRAMES to work for _REAL inputs
For _REAL the output cube was being created and mapped with type
_REAL but the individual 2d input images were being mapped with
type _DOUBLE by smf_open_file. This meant that the memcpy() was
failing badly. The quick workaround is just to force the output
cube to be _DOUBLE. Obviously if this was a KAPPA application
that would be very naughty.