smurf: Close heater efficiency file as soon as possible
The smfData with the heater efficiency data was associated with the
NDF. I was finding that for long map-maker runs we were getting
a problem at the end when the smfData was closed because the
efficiency data files are usually stored on an NFS mount point.

Now immediately copy the efficiency data and close the file.