smurf: Can not use PIXHEAT when assigning flats to science data
As of 20110901 we do a heater track immediately after a flat ramp
so there will normally be a slight change to PIXHEAT. The test
for usefulness was using PIXHEAT and determining that a failed
flat ramp didn't really matter. Now for science we do not include
PIXHEAT in the check.

The best solution is to only do a heater track at the start of
the flat ramp.