ast: Speed up plotting of polylines
Due to the recursive, adaptive way in which AST plots curves, each curve
can end up being plotted as hundreds or even thousands of tiny
disconnected polylines (depending on the value of the Tol attribute).
Each such polyline was previously drawn using a separate call to the
polyline plotting function in the underlying graphics system. This was
not much of a problem when using pgplot or other such compiled graphics
system. But for interpreted graphics systems such as Python's matplotlib,
it could be a crippling problem, speed-wise.

So this commit changes the way that polylines are drawn to avoid making
so many calls to the underlying polyline drawing function. All the
polylines are buffered, rather than being drawn immediately, and before
drawing them an attempt is made to reduce the number of polylines to be
drawn by looking for individual polylines that start and/or end at the
same point. Such polylines are combined into a single polyline, which is
then sent to the underlying polyline function.