smurf: first step towards iterative PCA: smf_clean_pca works on a chunk
Since PCA cleaning works better on small bits of data, yet we want to
reduce large data sets with a shared map iteratively, the approach I'm
going to attempt is to clean a concatenated time-series in many small
pieces. To do this, smf_clean_pca needs to be able to operate on
a sub-range of a full smfData. This is now accomplished with the
t_first and t_last parameters.