pcs: Allow the user to control messages through environment parameters.
Although the MSG_ tuning environment variables have been available and
documented for two decades they are in effective in most applications
because only one invoked MSG_TUNE( 'ENVIRONMENT'... and that was in
individual HTOOLS tasks.  Here I've added it to the fixed part.  The
main benefit is to be able obtain wider lines, say to report long NDF
paths through the MSG_SZOUT environment variable.  Some tasks may
smaller buffers than MSG_SZOUT and you might want to be able to adjust
to terminal width, but this is at least a step in the right direction.