smurf: new task SC2FILTERMAP with whitening as only available option
A slightly messy commit due to the need to get a new task written in
order to test and debug the 2d map filtering code. The only
map filter (other than the identity which does nothing) is a
whitening filter that measures a power law + white noise level
to the azimuthally averaged angular power spectrum. This power
spectrum is measured using a separate reference map, the idea
being that you measure it with a jackknife or something like that,
and then you apply it both to the science map and a beammap.

SC2FILTERMAP should eventually be able to do all of the filtering
type stuff that we currently do (like matched-filter). I need
to add other filter options, and 2d map apodization etc. first.

I also found two major bugs in the macro FFT_INDEX_TO_FREQ in
smf.h.source. I had some brackets in the wrong place, and I also
need to typecast everything as a signed data type since the FFT
has both positive and negative frequencies (even though I currently
convert everything into a scalar distance when I actually use it).
I'm astounded that sc2mapfft with azavpspec appeared to work.