surf: Add alternative skydip fitting routine based on GSL
This can replace the SCULIB_FIT_SKYDIP fortran code. SURF originally
used a NAG routine but this was replaced around 1995 when NAG was
pulled from the Starlink software. The replacement algorithm does
mostly work okay except that it will not fix fixed parameters. The
GSL replacement will leave fixed parameters fixed but still has
a problem with free parameters going out of range because the
GSL routine also does not have box constraints but uses a
penalty function to discourage bad parameter choices.

Note that there is a LM fitting routine in SURF which does
not seem to be used. The original NAG routine supported
non-linear constraints as well as box constraints.