smurf: cancel REF in smurf_mon instead of smf_getrefwcs
This parCancl in smf_getrefwcs seemed to be covering for some
sort of bug in par/ndf/hds. Since I later try to read REF
again inside smf_calcmodel_ast it was prompting for REF since
it was cancelled. Moving the cancel to smurf_mon gets it to run
without the prompt.