ndg: Plaster fix to enable NDF wildcarding to work again.
I may not have followed David's logic.  NDF wildcarding was failing
because ndg1_fpars was being called with NDOT=1, something the
documentation says should not occur for NDFs.  As a consequence, it
undergoes a second IDOT loop in ndg1_fpars where the suffix bounds are
lost as SUFBEG becomes -1 and a null suffix returned.  This in turn
leads to the name being placed in quotes and HDS_FIND returns bad
status because there's no SDF component.  Hence all the NDFs are
rejected.

I've made a crude fix as it's holding up my progress, and have not
checked that multiple-dot FITS files still work.  David can correct
to his intended logic.