smurf: Increase number of digits in DEFLON/DEFLAT string
The rounding in the default base position (DEFLON and DEFLAT) was
causing situations where the base position requested by the observer
was half an arcsec different to the base position reported by
SMURF. This usually doesn't matter but the way things are set
up in the pipeline and metadata table for the science archive
means that we need DEFLON and DEFLAT to be closer to BASEC1 and
BASEC2.

We now use a digits() attribute with value 8 for formatting the
base position to a string.