smurf: Add facility for dumping coords of every sample
Controlled by new exportlonlat config parameter (default 0).

For each input data file, longitude and latitude values (in degs, and
after downsampling) are stored in two separate 2D NDFs. Each new NDF has
the base name of the corresponding input NDF, but with an appended suffix
equal to the AST Symbol value for the relevant axis of the output
skyframe (e.g. "_ra", "_dec", "_glon", "_glat", etc). The pixel axes of
each NDF correspond to (ibolo,itime). The character components of each
NDF are set to indicate the contents of the NDF.