smurf: Multi-thread WVM calculation
The new WVM code can take quite a bit longer than the old code
so now we multi-thread the core WVM calculation. On my 4 core
machine for a test data set the run time went from 8.5 to 3.0
seconds.

API change to smf_correct_extinction to allow the worker thread
struct to be passed in. Side effect of this is that the EXTINCTION
task is now also multi-threaded.

This patch does not thread the WVM smoothing and nor does it thread
the extinction calculation itself. For my test data set the non-WVM
calculations took less than a second so threading won't save
much time.