smurf: Handle TCS excursion in 20110530#29
For this observation the telescope lost its mind at the end so the
positions in the data are a long way from the source. Attempting
to reconstruct a map leads to enormous map bounds. This patch
simply blanks the TCS data for the final 3 subscans of the observation
which effectively hides them from the map.