smurf: Only calculate WVM tau once for each chunk
smf_model_create was calling smf_correct_extinction for each subarray
which meant that we re-calculated the WVM data each time. Now the
first time smf_correct_extinction is called for a chunk we reuse
it for subsequent subarrays.

Note that there is a risk that the first subarray did not retrieve
a full set of WVM data and this would affect the remaining subarrays.