pal: Begin de-slalib-ification of Starlink C code
The PAL library is a new library designed to replace Fortran SLA
by providing a compatible C interface but written from
scratch in C. Some routines are rewrites of the Fortran code
whereas others are wrappers around SOFA routines.

This is the first commit and only a handful of routines have
been implemented.

The PAL name comes from "Positional Astronomy Library" as used
in the AST internals and in the Starlink Java PAL module. We
are deliberately not using the SLA name.