pal: Add palDmat
There may be better implementations available but this one
is a straight port of the Fortran and obfuscated C versions
with some attempt at sorting out variable scoping.

Neither of the original versions had comments and both may
as well have been obfuscated for all the clarity that exists
in the current code.