smurf: Flag constrained map pixels rather than zeroing them
Previously, smf_calcmodel_ast both flagged and set to zero any map
pixels that were constrained to zero by a mask. Now, masked map
pixels are still flagged (in the map quality array) but are no longer
set to zero.

The motivation for this is that the COM masking may need the unmodified
map to create another mask (the COM and AST masks may be different, e.g.
if COM.ZERO_SNR and AST.ZERO_SNR are set to different values). If
smf_calcmodel_ast blanks off the source area of the map by filling it
with zeros, then the subsequent call to smf_calcmodel_com on the next
iteration will not be able to calculate the SNR in those areas, and so
will not be able to create a COM mask.

The only place where the zero-filling was important was in smf_iteratemap
when the map is sampled to form the AST model that is then subtracted
from the residuals. This has been changed to check the quality of each
pixel, rather than relying on the pixel values being zero.