smurf: Use SOFA routines for smf_geod and smf_terr
Now uses iauGc2gd and iauGd2gc for conversion of geocentric to
geodetic coordinates. The smf_geod and smf_terr routines are now
very simple and exist to ensure that we use a consistent
reference ellipsoid in SMURF.

The change to WGS84 seems to move JCMT up 3 metres.

This patch adds SOFA as a dependency for SMURF.