smurf: Remove QLMAKEMAP
It has not served a useful purpose since about 2009 and makes maps
that look worse than the iterative map-maker in a longer time than
required by the iterative map-maker.

We could resurrect it as a wrapper around the iterative map-maker
that has a hard-wired set of config parameters but it doesn't really
seem worth it.