prm: Basic support for 64-bit integer types
PRM constants are now generated with INTEGER*8 values using a
"K" suffix (VAL__BADK etc). We now assume that modern compilers
can handle INTEGER*8 and int64_t data types.

This patch also includes a NUM_KTOI function for converting
64-bit to 32-bit integers. Other conversion routines will
need to be added as required. The NUM_KTOI function is implemented
in C as that seems to be more portable than relying on a test for
each fortran compiler.