GAIA: Add control of the standard of rest for cubes.
At present this does not validate that all the required attributes are
available for a particular standard of rest, and making that option
grey in the menu.

The standard of rest is not preserved in the history when using the Go
menu.
(cherry picked from commit 1f0ba53ce3458906709a420c107e27fa2ee4d8d3)