prm: generic conversion routines.
Needed a generic file for each input type. The output code is the
generic code since GENERIC can not handle <T1>TO<T2> conversions.

The BTOB, ITOI etc were handled in a previous commit hence the
strange names of the generic files since .gn would not remove
the type that matches the input.