smurf: Fix serious problem finding gains and offsets
Due to comparing status, rather than *status, with zero, the array of
flags indicating if each bolo-block needs to be fitted was being left
uninitialised, causing a random selection of bolo blocks to be rejected
on each iteration. This could account for some of the spurious behaviour
seen with the new COM model.