smurf: Deal with uninitialised SCUBA-2 header
Unfortunately a broken SCUBA-2 FITS header is a header filled with
zeroes rather than no header or a header with spaces. This meant
that we did end up with a header hash but it didn't have any
relevant cards in it. We now count to make sure we have cards
although we could force a check for INSTRUME.