smurf: Allow FLT model to be undone at start of iteration
New config parameter FLT.UNDOFIRST controls when the FLT model is
undone. The default value is 1 which means undo the FLT model at the
start of the iteration. This seems to give better convergence. If set
to zero, then the old behaviour is reinstated (i.e. the FLT model
is undone just before estimating a new FLT model in smf_calcmodel_flt).