smurf: Fix critical WVM indexing bug
When the WVM calculation was rewritten to be multi-threaded
a couple of indexes remained at their original values of 0 and
nframes in loops. This wasn't really obvious in the test data
I was using and with only 4 cores but with 32 threads and
occassional glitches the WVM trends towards the value of the last
threads to complete.