smurf: Ensure that WVM threads are split properly
There are a variable number of SCUBA-2 samples for every WVM
reading. In the past we naively split the time-series into equal
chunks even if that split WVM readings down the middle. This
had the disadvantage of calculating the same WVM measurement
multiple times whilst also giving an answer that changed depending
on the number of active threads. The reason for the latter being
that the airmass would be different at a different section of the
time slice but the WVM measurement is only valid for the start
of the value.

Now we scan through the WVM data looking for each new measurement
and then we decide where to split up the worker threads so that
they coincide with those boundaries. If there are fewer measurements
than workers we simply use fewer workers.

No attempt is made to correct for a phase shift in WVM_TIME
relative to RTS_END.
(cherry picked from commit b871a97cf4e70847992feb9774d7833cddf472b7)