irq: Assign defined value TRUE to component "VALID" in IRQ quality struture
Previously, VALID was left undefined since it's value is insignificant
(its mere existence, even with an undefined value, was a flag that the
structure was complete). To avoid potential problems with software that
expects all HDS component to have defined values, this has now been
changed so that VALUE is assinged the defined value "TRUE".