init: Add x86_64 starjava search path on OSX
The OS X library path test for starjava does not know whether we
are running a 64-bit java or a 32-bit java. Since OSX will reject
libraries of the wrong architecture we just add both to the
search path.

In normal usage this is not an issue with the java applications
because the class loader knows which architecture it is running
and sorts itself out. This patch addresses a problem in the
SPLAT build which requires JNIAST to be present to calculate the
line catalogues. SPLAT does not look in ../jniast so it fails
to find the x86_64 shared library locally. The linker then goes
through DYLD_LIBRARY_PATH looking for candidates and previously
it threw them out because only i386 was on the path.