ndf: New tuning parameter SECMAX gives max. no. of pixels in a section
An error is reported if an attempt is made to create an NDF section
containing more pixel than specified by SECMAX. The SECMAX value is given
in units of mega-pixels in order to allow use of existing INTEGER*4
tuning parameter system. The default value 2147 mega-pixels, which
corresponds to the maximum number of pixels that can be stored in an
INTEGER*4 (other things start going wrong for NDFs containing more than
2147 mega-pixels, for instance NDFTRACE reports it as having a negative
number of pixels).