ary: Fix bug causing incorrect data type when propagating compressed arrays
If NDF_PROP was used to copy a compressed input NDF to an output NDF,
without propagating the DATA array (i.e. the usual pattern), then the
output NDF would be given the data type of the array holding the
compressed values, rather than the data type of the uncompressed values.

For instance, if a _DOUBLE NDF is compressed into _WORD, and then used as
input to (say) kappa:cadd, the output from cadd would have data type
_WORD rather than _DOUBLE, causing havoc with the values stored in it.