Explicitly use X_CFLAGS
It turns out that IDI, GWM, GKS and CCDPACK were not explicitly
adding the X11 includes path to the compiler options. This was noticed
when OS X Mountain Lion stopped adding /usr/X11/include to the build
automatically. We now add $X_CLAGS to AM_CFLAGS or equivalent.

GKS was also tweaked to stop the build early if X11 was missing
rather than stopping the build when the xwindows driver was being
built. This was partly driven by the global GKS configure.ac not
adding X_FLAGS to includes before attempting to check whether X11
includes were available.