docs: Update email address in doc subdir
I have not attempted to rewrite documents and the
documents in the obsolete directory have not been
modified at all. configure.ac and component.xml have
been updated and only a few cases in documents that
refer to the Starlink Libraries contact email.
This has not updated the obsolete documentation itsel