fluxes: Remove JPL dependency
FLUXES didn't really need to subarcsecond accuracy for planet
positions provided by the JPL ephemeris so we remove the
complication. We now use SLALIB to calculate the positions and
take the hit of up to 1 arcmin in error. This seems to translate
to no more than 0.01% error in flux for some planets.