fluxes: Simplify environment variables
FLUXES disappears and is replaced with FLUXES_DIR that is already
being set by the Starlink login.

FLUXPWD is now optional and if it is missing (which it will be
in Starlink) we use PSX_GETCWD to provide the default full path
to the current directory.
(cherry picked from commit d3476d9ac9c16f2506011c604bf269131db7dc76)