ast: New astRebinSeq flag AST__NONORM to indicate no normalisation needed
In previous versiosn of AST, a null pointer could be provided for the
astRebinSeq weights array so long as the AST__REBINEND flag was never
used. This is useful if you do not want to normalise the final output
arrays (i.e. if you want the simple sum of all input data within each
output pixel). This was an undocumented feature which I forgot about when
making the recent changes to astRebinSeq. This commit re-instates the
feature but formalises it by requiring a documented flag (AST__NONORM) to
be supplied if no weights are being used.
(cherry picked from commit 54c33d211ce46b20ae2e1ea8ab7c8d3338b580f3)