smurf: New version of POL-2 bonus POL_ANG fixing code
The old version found lots of bonus points because it did not correct
later data as soon as it found a bonus point and therefore got confused
by the long RTS_END gap every 25th sample. The new version is different
to the example python code written by Ryan in that it does not assume a
constant rotation speed, and does not need to be told what the
nominal rotation speed is.