docs: Preliminary version of SCUBA-2 Cookbook for Kapuahi.
We need a version of the SCUBA-2 Cookbook in Kapuahi to complement the
release of much-improved SCUBA-2 reduction software.  Now this
incarnation is incomplete, has yet to be proof-read and have its
spelling checked, still has commented text, and does not build.  It
will be refined in further versions.
(cherry picked from commit 634605d02194a282c2665658ca009ef77a408708)