gsd: Define size of name buffer in a consistent manner
This fixes the very long standing segv we were getting in the
perl GSD test suite with gsdGet1r. We used to get it on linux
and it would be fixed if we turned down optimisation levels. We
therefore put it down to a compiler bug. Today I started getting
it from clang on OSX and it was hitting a stack corruption error.
Looking at the only variable allocated on the stack in gsdGet1x
I had:

   char name[15];

and of course that is too small for a string of length 15 plus
a nul terminator. Changing that to 16 fixed everything but I
took the opportunity to use a constant for it throughout the
library.
(cherry picked from commit 04ac25e299557d6eb247bbd01b2686797d484513)