docs: SC/21---Introduce PNG graphics for hypertext.
This is for three reasons.  The PostScript files are clipped in the
browser, are of poorer quality, and seem to cause latex2html to mislocate
the images (the figures, symbols, and formulae).  They are too large
at present, especially for pairs that are supposed to reside side by
side.  I could delay further saving storage, but committing now offers
a chance for other eyes to see what may be causing latex2html to
misbehave.  These have helped, but there are still lots of misplaced
images.