docs: SC/21---Use text links for hypertext for internal cross referencing.
Those boxes for internal cross-referencing look naff especially the
cluster of them in the Introduction, and they aren't as clear as text
strings.  Added a newcommand to simplify this.  In some cases this was
not ideal; to avoid repeating the title in the link as already written
or paraphrased in the sentence there are a few "this appendix" or
equivalent.

There was a bonus in that some named objects previously did not begin
with a capital letter and were missing the ~ to prevent an object and
its \ref{label} number from being broken across lines.
(cherry picked from commit 777836c20122fbe7b8a6e2d577b5d7386127460f)