smurf: Handle case where we have a short flat ramp
If the ramp is truncated it is possible to only get 3 measurements
and this would fail in smf_stats1I. We now trap this situation
and calculate a simple mean and a grossly inaccurate standard
deviation so that we can have data for the fit.

Also simplified the counting to check for bad values as we extract
the heater data.