smurf: Correct usage of astRebinSeq flags when rebinning an acsis cube
The final call to astRebinSeq within smf_rebincube_ast, which normalizes
the output NDF data and variance values, was failing to pass the
AST__USEVAR flag to astRebinSeq. Changes in AST now mean that this causes
all the output variances to be set bad. So now make sure that the
AST_VARWGT flag is passed to astRebinSeq if required so that good output
variances are generated.