kaplibs: Allow DEVICE=! to be for looping graphics applications
Previously, running "normalize $KAPPA_DIR/* device=!" would attempt to
prompt for a new graphics device when plotting the second normalisation
plot. This is because the initial null value for DEVICE generated an
error in KPG1_GRAPH when closing down PGPLOT, resulting in the parameter
value being cancelled in KPG1_PGCLS, resulting in a user prompt when the
DEVICE parameter was next accessed to display the normalisation plot on
the next loop.