smurf: Allow initial sky to be specified when running makemap
To use this, set the config parameter IMPORTKY to indicate the name of
the ADAM parameter that should be used to access the initial sky
estimate. The sky map is sampled and subtracted from the cleaned time
series data before the first iteration starts. It is also added back in
before the new map is made at the end of the first iteration.