specx: Fix formatted read issue on g95 and gfortran
It seems that

      FORMAT(I2,'/',I2,'/',I2)

is not allowed in the Fortran standard. Instead this has to be
written as:

      FORMAT(I2,1X,I2,1X,I2)

Allegedly Sun and Intel compilers do that transparently and the
fixed string in the format was never actually looked at when reading
the numbers.

This fixes the runtime gfortran error:

 >> read-fits-spectrum
 ILEN =           23
At line 139 of file parse_date_obs.f
Fortran runtime error: Constant string in input format
(I4,'-',I2,'-',I2,'T',I2,':',I2,':',I2,'.',I3)