specx: A serious attempt at porting SPECX to 64-bit
SPECX now runs up on 64-bit OSX, can read spectra and maps, can
display channel maps and normal maps. All pointer access is now
mediated by CNF.

 - %VAL calls now use CNF_PVAL
 - %LOC calls now use CNF_PREG

EXTREME was used for most of the %VAL conversion so there is no
commentary in the prologues of modified files.

Unfortunately there are bound to be more bugs in this version than
in the pure 32-bit implementation. The main issue concerns pointer
arithmetic. There is quite a lot of code that does

  LOCATION = PNTR + OFFSET
  CALL FUNCTION( %VAL( LOCATION ) )

which worked fine on 32-bit but for the naive port of:

  LOCATION = PNTR + OFFSET
  CALL FUNCTION( %VAL( CNF_PVAL( LOCATION ) ) )

things now go boom on 64-bit *and* 32-bit. The problem being
that CNF_PVAL doesn't know anything about the compressed pointer
at LOCATION. The fix is

  CALL FUNCTION( %VAL( CNF_PVAL( PNTR ) + OFFSET ) )

I've fixed all the ones I could find but specx is now liable to go
boom for the cases I've missed. All we can do is ask people to test
it on 64-bit and delay updating the 32-bit version built at JAC.