kappa: It is not always correct for WCSADD to transfer Frame attributes
So the new boolean parameter TRANSFER can now be used to inhibit the
transfer of set variables from basis to new Frame. The dynamic default is
to transfer them only if the two Frames have the same class and Domain.

It is just about possible that this changed behaviour may cause some
scripts to break. But I cannot easily think of any situation where you
would want to transfer attribute values from a Frame of a different class
or Domain...