smurf: algorithm is OK; clean up comments and whitespace
In 0d718a8f0b4be114ba98a67d39384fd3dcfa87f8 I added a comment that I
thought we should divide by the complex conjugate of the filter, even
though tests suggested otherwise. I now understand we divide directly by
the filter: according to the IDL code on which this routine is based,
they had to reverse the order of the inverse FFT in order to obtain the
impulse response (i.e., this filter code really gives the fourier space
coefficients of the filter itself, not the FFT of the impulse response).
So, I've removed the confusing comment...