smurf: Provide config parameter to control MCE filtering of fakemap data
New parameter "fakemce" defaults to zero. If set to one, the raw time
series data is smoothed using the MCE response *AFTER* the fakemap data
has been added to it. Ideally, the fake time-series data should be
smoothed *before* adding it to the real time-series data, but this
requires more work.