ast: Use FITS compliant CTYPE for unknown spherical coordinate axes
When exporting a SkyFrame with "System=UNKNOWN" to FITS-WCS, use the
SkyFrame axis symbols for chars 1-4 of CTYPE if they are of the form
"xxLN/xxLT" or "xLON/xLAT". Otherwise use "UNLN/UNLT". This ensures that
a SkyFrame with "System=UNKNOWN" will be used to describe the WCS when it
is read back in again.