smurf: Retain the previous converged status of a map on the final iteration
If a map that uses a mask converges, it does one last iteration to allow
the masked area to be set to non-zero values. If this last iteration
produced a mean normalised change in the map that is larger than maptol,
then previously a message would be displayed saying that the map did not
converge, even though it did converge (prior to the final iteration).
Jess does not like this, so I've changed it so it no longer changes the
converged status on the final iteration.