kappa: Update version and documentation for recent CONFIGECHO changes.
I had already incremented to 2.0-9 for David's new DEFVAL parameter,
but forgot to commit before going to CADC.  So I've merged all the
new parameters into one increment.

Please note the Examples in SUN/95 are consistently are what PCS
expects, which is not the same as the UNIX shell.  Also named/titled
parameters start with a capital, because it's a name.  So I've undone
many of DSB's changes which in turn had undone my previous editorial
work and Graham's correct additions.

I also compressed the <***> for html to prevent spaces appearing around
each character.