smurf: Recalculate WVM tau in output headers
The WVMTAUST and WVMTAUEN headers are calculated using the current
conversion values at the summit when the data are taken. These
values are usually incorrect as we have improved the WVM calculation
significantly over time. To make it more obvious which values are
being used for extinction correction internally we now recalculate
the WVM tau and update the relevant FITS headers.

This patch does not update the associated WVM times.