agi: Allow EPS output to be merged with previous EPS output
When using native PGPLOT, AGI now provides additional Postscript devices
prefexed with "A" ("/APS", "/ACPS", etc). These devices get translated
into the standard device names ("/PS", "/CPS", etc) before PGPLOT is
opened, but AGI arranges for the Postscipt output created by PGPLOT to be
merged automatically with any pre-existing Postscript file with the same
name. This does away with all that tedious messing about with psmerge...

Combined with a Postscript viewer such as "okular", which re-draws the
display if the on-disk file changes, this provides a similar experience
to using a persistent X window, with the benfit that you end up with a
postscript file to include in your latex document rather than a low-res
screen dump.

Jess watch out...