smurf: Add fts_port parameter to sc2ast functions.
This parameter indicates whether to include the FTS-2 features
in the returned frameset.  If set to a true value (not NO_FTS
which is defined as zero), then the FTS-2 code is activated.
The parameter currently has options for the tracking and
image ports.  This could be made to select the current one of
two FTS-2 variants, but for now, and as agreed in last week's
meeting, it simply controls which single mapping to sky
coordinates is required.