smurf: WCS for FTS-2 now includes two variant Mappings
If FTS-2 is in the beam, sc2ast now return a FrameSet containing two
variant Mappings for the FPLANE and SKY Frames - one for each FTS port.
The variant mappings are labelled PORT1 and PORT2, and can be selected by
setting the "Variant" attribute of the FrameSet to the name of the
required variant (you need to make SKY or FPLANE the current Frame
first). This has required a small change to the structure of the FrameSet
- the SKY Frame is now derived from the FPLANE Frame, rather than the
GRID Frame as before. This causes the mapping transformations to be
applied in a different order. The old and new orders are mathematically
equivalent, but produce different numerical rounding errors. Consequently
some samples that are close to the edge of map pixels will move into
adjacent pixels, resulting in some small changes to the resulting map.