Add external C code for reading new format CSO fit data files
These files were developed outside of SMURF for reading CSO fit data
files. The data files use the new format developed for SCUBA-2 rather
than the old format used for SCUBA and defined in the JCMT::Tau::CsoFit
perl module. Hence the use of the "2" in the names.

An example fit data file is included covering 20130313 to 20130501.