smurf: Initialise "dat" even when status good
With commit dbd9e70bae48b9327106687bf7872ba429689ffe we started freeing
resources inside the "dat" structure. This caused problems if status
was set to bad early on because the "dat" struct itself was only being
initialised if status was good. This patch moves the memset outside
the status check.