smurf: Better gap filling
Because of the size of the fitting box, the old scheme could not handle
lots of small sections (less than 50 samples) of good and bad values next
to each other. The new scheme uses several passes. The first pass is just
like it used tobe. If the first pass leaves any unfilled gapos, then a
second pass is performed with a smaller fitting box (25 samples). If this
leaves any unfilled gaps a third pass is performed with an even smaller
fitting box (12 samples). If this leaves any gaps, an error is reported.

This issue cropped up making a map from FTS-2 data. The symptom was an
error in kpg1Fit1D.