smurf: Correctly reorder the GAI model when required
This was potentially a very serious bug - requests to reorder the GAI
model were incorrectly causing the residuals to be reordered instead.
However, the order of the GAI model is stipulated only in one place -
smf_calcmodel_com - and consequently no re-ordering of the GAI model is
ever actually needed. So in fact this bug fix should have no effect.