smurf: Merge BEGIN and END FITS headers as floats
When the BEGIN and END FITS headers are copied between two
headers before merging, smf__fits_copy_items now checks
if the value is of floating point type and if so copies
it as such.  This is intended to prevent the problem of
FITS headers such as WVMTAUEN vanishing if after converting
to a string and back to a floating point number and back
to a string again the string isn't the same as the string
to which the original value converts.