smurf: Disable WVM corrections since 20130409
Jessica Dempsey requested that the WVM corrections be
removed for now since the "Silver" WVM does not show an
obvious deviation from the CSO Tau meter on the
nightly Tau plots.

Change applied from 20130409 as the performance shifting
"Silver" WVM was replaced with the "Black" WVM on this
date, and where comparison is reasonable, the "Black"
WVM did not shown an obvious deviation from the CSO
Tau meter either.

Also attempted to make the date ranges a bit clearer
in the source code.