KAPPA: Permits more control of beam size and shape in BEAMFIT.
Note there is a change to the data type and function of Parameter
FIXFWHM.  It's now _LOGICAL.  Parameter FWHM is used to set fixed or
initial guesses for the full-width half maxima.  There is also a
new CIRCULAR logical parameter which works in conjunction with
Parameter FWHM.

[I wonder if there should be further such separations, such as
a FIXAX parameter to go with AXRATIO so that AXRATIO either sets
an initial guess or a fixed ratio depending on FIXAX, and FIXBACK
plus BACK.  It might be shorter to have FIX=<comma-separated list>
with abbreviations".]

For circular beams the the output should now not list the minor axis
and orientation.