smurf: Enable new tausrc AUTO mode
Now that we have CSO fits available AUTO mode has been refactored
to first look for the WVM and then look to CSO fits. If neither
WVM nor CSO fits are available extinction is not possible and
an error is raised.

This patch lead to some minor refactoring to change the way that
thw wvm time-series cache is checked. We assume that it's always
going to be used if we are in AUTO, WVM or CSOFIT mode rather than
checking each mode separately.